// apps/admin/src/modules/history/types.ts
// Copyright LEAV Solutions 2017 until 2023/11/05, Copyright Aristid from 2023/11/06
// This file is released under LGPL V3
// License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
export type HistoryData = {
    key: string;
    date: string;
    user: string;
    action: string;
    object: string;
    entity: string;
    details: string;
    before: string;
    after: string;
    queryId: string;
    rawJson: string;
};